calling_convention: refactor repository structure
