+++ include/retdec/config/calling_convention.h
@@ -97,14 +97,11 @@ class CallingConvention
 			CC_SPECIALE,
 			CC_SPECIALP,
 			CC_SPECIAL,
-			CC_FASTCALL_PASCAL,
 			CC_WATCOM,
-			CC_SYSTEMVX64,
-			CC_MICROSOFTX64,
+			CC_X64,
 			CC_ARM,
 			CC_ARM64,
 			CC_MIPS,
-			CC_MIPSPSP,
 			CC_MIPS64,
 			CC_POWERPC,
 			CC_POWERPC64,
