+++ src/bin2llvmir/providers/abi/mips.cpp
@@ -28,14 +28,7 @@ AbiMips::AbiMips(llvm::Module* m, Config* c) :
 			MIPS_REG_A2,
 			MIPS_REG_A3};
 
-	if (c->getConfig().tools.isPspGcc())
-	{
-		_defcc = CallingConvention::ID::CC_MIPSPSP;
-	}
-	else
-	{
-		_defcc = CallingConvention::ID::CC_MIPS;
-	}
+	_defcc = CallingConvention::ID::CC_MIPS;
 }
 
 AbiMips::~AbiMips()
