+++ src/bin2llvmir/providers/abi/x64.cpp
@@ -29,7 +29,7 @@ AbiX64::AbiX64(llvm::Module* m, Config* c) :
 		X86_REG_R8,
 		X86_REG_R9};
 
-	_defcc = CallingConvention::ID::CC_SYSTEMVX64;
+	_defcc = CallingConvention::ID::CC_X64;
 }
 
 AbiX64::~AbiX64()
