+++ src/bin2llvmir/providers/calling_convention/pic32/pic32_conv.cpp
@@ -1,10 +1,11 @@
 /**
- * @file src/bin2llvmir/providers/calling_convention/pic32.cpp
+ * @file src/bin2llvmir/providers/calling_convention/pic32/pic32_conv.cpp
  * @brief Calling conventions of PIC32 architecture.
  * @copyright (c) 2019 Avast Software, licensed under the MIT license
  */
 
-#include "retdec/bin2llvmir/providers/calling_convention/pic32.h"
+#include "retdec/bin2llvmir/providers/abi/abi.h"
+#include "retdec/bin2llvmir/providers/calling_convention/pic32/pic32_conv.h"
 #include "retdec/capstone2llvmir/mips/mips.h"
 
 namespace retdec {
@@ -35,6 +36,11 @@ Pic32CallingConvention::~Pic32CallingConvention()
 
 CallingConvention::Ptr Pic32CallingConvention::create(const Abi* a)
 {
+	if (!a->isPic32())
+	{
+		return nullptr;
+	}
+
 	return std::make_unique<Pic32CallingConvention>(a);
 }
